Add slur

+++ espnet2/svs/xiaoice/XiaoiceSing.py
@@ -450,7 +450,7 @@ class XiaoiceSing(AbsSVS):
                 value (LongTensor): Batch of padded duration (B, Tmax).
             duration_length (Optional[Dict]): key is "lab", "score_phn" or "score_syb";
                 value (LongTensor): Batch of the lengths of padded duration (B, ).
-            slur (FloatTensor): Batch of padded slur (B, Tmax).
+            slur (LongTensor): Batch of padded slur (B, Tmax).
             slur_lengths (LongTensor): Batch of the lengths of padded slur (B, ).
             spembs (Optional[Tensor]): Batch of speaker embeddings (B, spk_embed_dim).
             sids (Optional[Tensor]): Batch of speaker IDs (B, 1).
